Create OddorEven.java
1. Design an algorithm that accepts two integer input from
the user. If the 2 inputs are the same, the program will tell
the user that the number he entered is equal. Else, using
loop, the program will print the numbers starting from the
first input until it reaches the second input and classify
the odd numbers from even numbers. Implement your
algorithm design in any programming language.